requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.3.0
yt-dlp>=2024.1.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
openai-whisper>=20240314